+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/vue`:
     -   Create the `Chip` component
+    -   Create the `ChipGroup` component
     -   Create the `Popover` component
     -   Create the `PopoverDialog` component
 -   `@lumx/react`:
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Improve spacing util size type (abstract sizes, not t-shirt sizes)
     -   Add all direction short-hand for spacing utils (`margin(null)`, `padding('regular')`, etc.)
     -   Moved `Chip` from `@lumx/react`
+    -   Moved `ChipGroup` from `@lumx/react`
     -   Moved `Popover` from `@lumx/react`
     -   Moved `PopoverDialog` from `@lumx/react`
 -   `@lumx/react`: